fix: seperate tracing and main modules
