+++ pkg/tracing/main.go
@@ -6,64 +6,9 @@ import (
 	"log"
 	"net/http"
 	"os"
-	"strconv"
 
-	"go.opentelemetry.io/otel"
-	"go.opentelemetry.io/otel/attribute"
-	"go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracehttp"
-	"go.opentelemetry.io/otel/sdk/resource"
-	sdktrace "go.opentelemetry.io/otel/sdk/trace"
-	semconv "go.opentelemetry.io/otel/semconv/v1.21.0"
-	"go.opentelemetry.io/otel/trace"
+	"github.com/informatik-mannheim/cmg-ss2025/pkg/tracing/tracing"
 )
-
-var tracer trace.Tracer = otel.Tracer("default-tracer")
-
-// Init sets up tracing with OTLP (Jaeger via OTLP collector endpoint)
-func Init(serviceName, otlpEndpoint string) (func(context.Context) error, error) {
-	ctx := context.Background()
-
-	exp, err := otlptracehttp.New(ctx,
-		otlptracehttp.WithEndpoint(otlpEndpoint),
-		otlptracehttp.WithInsecure(), // I would say this is fine because it should be in a local network, so no TLS is needed
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	res, err := resource.New(ctx,
-		resource.WithAttributes(
-			semconv.ServiceNameKey.String(serviceName),
-		),
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	tp := sdktrace.NewTracerProvider(
-		sdktrace.WithBatcher(exp),
-		sdktrace.WithResource(res),
-	)
-	otel.SetTracerProvider(tp)
-	tracer = tp.Tracer(serviceName)
-
-	return tp.Shutdown, nil
-}
-
-func Middleware(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		ctx, span := tracer.Start(r.Context(), r.URL.Path)
-		defer span.End()
-
-		span.SetAttributes(
-			attribute.String("http.method", r.Method),
-			attribute.String("http.url", r.URL.String()),
-			attribute.String("http.code", strconv.Itoa(r.Response.StatusCode)),
-		)
-
-		next.ServeHTTP(w, r.WithContext(ctx))
-	})
-}
 
 // Just an example main function to test the tracing setup
 func main() {
@@ -73,7 +18,7 @@ func main() {
 	}
 
 	// Replace with your actual Jaeger endpoint!
-	shutdown, err := Init("test-service", jaeger)
+	shutdown, err := tracing.Init("test-service", jaeger)
 	if err != nil {
 		log.Fatal("Tracing init failed:", err)
 	}
@@ -85,5 +30,5 @@ func main() {
 	})
 
 	log.Println("Listening on :8080...")
-	http.ListenAndServe(":8080", Middleware(mux))
+	http.ListenAndServe(":8080", tracing.Middleware(mux))
 }
